requirements.txt
aiogram==3.13.1
pyrogram==2.0.106
tgcrypto==1.2.5
gspread==6.1.2
google-auth==2.29.0
pytz==2024.1
openpyxl==3.1.2
python-dotenv==1.0.1
